activity email refactorings WIP
