+++ kifi-backend/modules/shoebox/app/com/keepit/commanders/emails/activity/UnreadMessageCountComponent.scala
@@ -1,0 +1,5 @@
+package com.keepit.commanders.emails.activity
+
+class UnreadMessageCountComponent {
+
+}
